requirements.txt
kucoin-futures-python